+++ client/src/components/paginaion/style.css
@@ -1,0 +1,147 @@
+.container-classified {
+  margin-right: 15px;
+  display: flex;
+  flex-direction: column;
+  margin: auto;
+}
+.sub-container {
+  display: flex;
+  flex-direction: row;
+  margin: 15px;
+  background: #ffffff;
+  border: 1px solid rgba(169, 175, 176, 0.5);
+  box-sizing: border-box;
+  border-radius: 8px;
+  text-decoration: none;
+}
+.rihgt {
+  width: 25%;
+  height: auto;
+  background-position: center;
+  background-size: cover;
+}
+
+.left {
+  width: 70%;
+  margin-right: 20px;
+}
+.class-img {
+  width: 100%;
+  height: 100%;
+  border-radius: 10px;
+}
+.favarite {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
+}
+.name-classified {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
+}
+.sub-name-classified {
+  font-weight: 600;
+  font-size: 16px;
+  line-height: 34px;
+  text-align: right;
+  color: #a9afb0;
+}
+.title-classified {
+  font-family: Cairo;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 20px;
+  color: #333333;
+  margin-bottom: 0;
+  margin-top: 2px;
+}
+
+.more-detalies {
+  display: flex;
+  flex-direction: column;
+  width: 50%;
+  justify-content: space-between;
+}
+.info {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin-top: 10px;
+}
+
+.icon-city {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  font-size: 20px;
+  text-align: right;
+  color: #aeb0a9;
+  margin: 0;
+}
+.sub-city {
+  color: #1a6e9a;
+  margin-right: 5px;
+}
+.price {
+  display: flex;
+  flex-direction: row;
+  font-size: 20px;
+  text-align: right;
+  color: #a9afb0;
+  margin: 0;
+}
+.sub-price {
+  color: #1a6e9a;
+  margin-right: 5px;
+}
+.title-decription {
+  font-weight: 600;
+  font-size: 24px;
+  line-height: 150%;
+  text-align: right;
+  color: #333333;
+}
+
+.section-btn-classified {
+  width: 350px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin-top: 10px;
+  margin-bottom: 10px;
+}
+
+.section-btn-classified button {
+  display: flex;
+  align-items: center;
+  width: 160px;
+  height: 38px;
+  text-align: center;
+  border-radius: 5px;
+  font-size: 18px;
+  line-height: 150%;
+  display: flex;
+  align-items: center;
+  text-align: center;
+  cursor: pointer;
+}
+
+.btn-chat-classified {
+  background-color: #1a6e9a;
+  color: #ffffff;
+  border: 1px solid #ffffff;
+}
+
+.btn-call-classified {
+  background-color: #ffffff;
+  color: #1a6e9a;
+  border: 1px solid #1a6e9a;
+}
+.btn-like-classified {
+  border: none;
+  background-color: #ffffff;
+  cursor: pointer;
+}
